requirements.txt
# Django and Core
Django==5.2.8
python-dotenv==1.2.1
psycopg[binary]==3.2.12
dj-database-url==2.1.0

# Django REST Framework
djangorestframework==3.16.1
django-cors-headers==4.9.0

# Authentication
djoser==2.3.3
djangorestframework-simplejwt==5.5.1
dj-rest-auth[with_social]==5.0.2
django-allauth==0.57.0

# API Documentation
drf-spectacular>=0.27.0

# drf_api_logger packages
drf_api_logger==1.1.21

# WhiteNoise packages
whitenoise==6.11.0

# Development and Testing
pytest==9.0.1
pytest-django==4.11.1
